Changes HunspellDictionary to an interface in the hunspell package and moves logic to AbstractHunspellDictionary.
